requirements-dev.txt
django-debug-toolbar>=1.6
flake8>=2.2.5
ipython>=5.0
isort>=4.2
pytest>=3.0
pytest-cov>=1.8
pytest-django>=2.8
pytest-spec>=0.2
pytest-sugar>=0.7
tox>=1.7
